vtcol.rs: choose correct type of argument passed to write()
D’oh!

Thanks to /u/DroidLogician for the pointer:
http://www.reddit.com/r/rust/comments/3588d4/suggestions_for_handling_architecturedependent/cr22okf

+++ src/vtcol.rs
@@ -601,7 +601,7 @@ get_console_fd
 fn
 write_to_term (fd : Fd, buf : &str)
 {
-    let len = buf.len() as u64;
+    let len = buf.len() as libc::size_t;
     let raw = std::ffi::CString::new(buf.as_bytes()).unwrap();
     unsafe { libc::write(fd, raw.as_ptr() as *const libc::c_void, len) };
 }
